pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "refund-guard"
version = "0.3.0"
description = "Turn a real order into a safe refund tool for your AI agent."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Matt Messinger" }]
keywords = ["ai", "agents", "refunds", "guardrails", "stripe", "safety"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[project.scripts]
refund-guard = "refund_guard.doctor:main"

[project.urls]
Homepage = "https://github.com/MattMessinger1/agentic_refund_guardrail"
Issues = "https://github.com/MattMessinger1/agentic_refund_guardrail/issues"
npm = "https://www.npmjs.com/package/@mattmessinger/refund-guard"

[tool.setuptools.packages.find]
include = ["refund_guard*"]
